RRDHOSTs free all their memory (all substructures) when they are deallocated
